reviews displaying with other reviews and on own individual pages

+++ src/main/java/com/filmreview/repositories/FilmReviewRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.filmreview.models.FilmReview;
 
-public interface FilmReviewRepo extends JpaRepository<FilmReview, Short>{
-
+public interface FilmReviewRepo extends JpaRepository<FilmReview, Long>{
+	
 }
